fix: 修复所有 clippy 警告，满足 GitHub Actions CI 要求
主要修复：
- 修复未使用的变量和导入
- 使用 base64 新 API (Engine::decode)
- 修复 vec-init-then-push, needless-borrow 等简单问题
- 修复 manual-range-contains, or_insert_with -> or_default
- 修复 matches! 宏使用
- 修复 clamp 函数使用
- 修复 filter().next() -> find()
- 修复 useless-vec, unnecessary-cast等issue

影响的包：
- browerai-deobfuscation (8 个文件)
- browerai-learning, browerai-persistent-layer
- browerai-redis-integration (3 个文件)
- browerai-renderer, browerai-ai-integration
- browerai-api-server, browerai (main)

全部通过: cargo clippy --workspace -- -D warnings ✅

+++ crates/browerai-learning/src/comparison_feedback.rs
@@ -96,7 +96,7 @@ impl ComparisonFeedbackGenerator {
 
         // 计算学习置信度
         let confidence =
-            (verification.verification_score * 0.4 + comparison.overall_similarity * 0.6) as f64;
+            verification.verification_score * 0.4 + comparison.overall_similarity * 0.6;
 
         Ok(ComparisonFeedback {
             verification_feedback: verification_fb,
